Switch from RESTeasy to Spring REST
And some cleanup as well
