+++ src/main/java/org/esbtools/gateway/resync/ResyncGatewayConfiguration.java
@@ -1,4 +1,4 @@
-package org.esbtools.gateway.resync.config;
+package org.esbtools.gateway.resync;
 
 import java.util.List;
