+++ src/main/webapp/WEB-INF/web.xml
@@ -1,18 +1,6 @@
-<web-app id="esb-resync-gateway" version="2.4"
-	xmlns="http://java.sun.com/xml/ns/j2ee"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://java.sun.com/xml/ns/j2ee
-	http://java.sun.com/xml/ns/j2ee/web-app_2_4.xsd">
+<web-app id="esb-resync-gateway" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://java.sun.com/xml/ns/javaee" xmlns:web="http://java.sun.com/xml/ns/javaee/web-app_2_5.xsd" xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_2_5.xsd" version="2.5">
 
 	<display-name>esb-resync-gateway</display-name>
-
-	<listener>
-		<listener-class>org.jboss.resteasy.plugins.server.servlet.ResteasyBootstrap</listener-class>
-	</listener>
-
-	<listener>
-		<listener-class>org.jboss.resteasy.plugins.spring.SpringContextLoaderListener</listener-class>
-	</listener>
 
 	<context-param>
 		<param-name>contextConfigLocation</param-name>
@@ -22,16 +10,15 @@
 	</context-param>
 
 	<servlet>
-		<servlet-name>resteasy-servlet</servlet-name>
-		<servlet-class>org.jboss.resteasy.plugins.server.servlet.HttpServletDispatcher</servlet-class>
-		<init-param>
-			<param-name>javax.ws.rs.Application</param-name>
-			<param-value>org.esbtools.gateway.resync.rest.ResyncGatewayApplication</param-value>
-		</init-param>
+		<servlet-name>rest</servlet-name>
+		<servlet-class>
+			org.springframework.web.servlet.DispatcherServlet
+		</servlet-class>
+		<load-on-startup>1</load-on-startup>
 	</servlet>
 
 	<servlet-mapping>
-		<servlet-name>resteasy-servlet</servlet-name>
+		<servlet-name>rest</servlet-name>
 		<url-pattern>/*</url-pattern>
 	</servlet-mapping>
