+++ src/test/java/org/esbtools/gateway/resync/ResyncGatewayConfigurationTest.java
@@ -1,4 +1,4 @@
-package org.esbtools.gateway.resync.config;
+package org.esbtools.gateway.resync;
 
 import org.junit.After;
 import org.junit.Before;
